+++ src/accessory/abstract/helper.ts
@@ -20,6 +20,8 @@ import { LightbulbAccessory } from '../onoff/lightbulb.js';
 import { OutletAccessory } from '../onoff/outlet.js';
 import { SwitchAccessory } from '../onoff/switch.js';
 
+import { BlindAccessory } from '../position/blind.js';
+
 import { AirSensorAccessory } from '../sensor/air.js';
 import { CO2SensorAccessory } from '../sensor/carbonDioxide.js';
 import { COSensorAccessory } from '../sensor/carbonMonoxide.js';
@@ -55,6 +57,8 @@ export function createAccessory(
     return new PurifierAccessory(Service, Characteristic, platformAccessory, accessoryConfig as Configs.PurifierConfig, log, isGrouped);
   case AccessoryType.AirQualitySensor:
     return new AirSensorAccessory(Service, Characteristic, platformAccessory, accessoryConfig as Configs.AirSensorConfig, log, isGrouped);
+  case AccessoryType.WindowCovering:
+    return new BlindAccessory(Service, Characteristic, platformAccessory, accessoryConfig as Configs.BlindConfig, log, isGrouped);
   case AccessoryType.CarbonDioxideSensor:
     return new CO2SensorAccessory(Service, Characteristic, platformAccessory, accessoryConfig as Configs.CO2SensorConfig, log, isGrouped);
   case AccessoryType.CarbonMonoxideSensor: