+++ CHANGELOG.md
@@ -2,10 +2,18 @@
 
 All notable changes to homebridge-dummy will be documented in this file.
 
-## 1.4.0-beta.16 (2025-09-26)
+## 1.4.0-beta.18 (2025-09-30)
 
 ### Added
-- `Air Purifier`, `Air Quality Sensor`, `Fanv2`, `GarageDoorOpener`, `HeaterCooler`, `LightSensor`, and `Valve`
+- A bunch of new accessory types:
+    - `Air Purifier`
+    - `Air Quality Sensor`
+    - `Fanv2`
+    - `GarageDoorOpener`
+    - `HeaterCooler`
+    - `LightSensor`
+    - `Valve`
+    - `Window Covering`
 - Broker, username, and password can be set via environment variables ([documentation](https://github.com/mpatfield/homebridge-easy-mqtt#environment-variables))
 - Publish arbitrary MQTT messages on connect ([documentation] (https://github.com/mpatfield/homebridge-easy-mqtt#mqtt-onconnect))
 - Traduceri în limba română. Mulțumesc, [@rursache](https://github.com/sponsors/rursache)!